ImageEncoder/requirements.txt
bitarray==3.1.1
Pillow==7.0.0
